refactor(ble): split handler-node-ble.ts into focused module directory
Mirror the #131 split applied to handler-mqtt-proxy. The 1186-line
src/ble/handler-node-ble.ts becomes src/ble/handler-node-ble/ with 9
modules (avg 139 lines each) so each concern lives in its own file:

* dbus.ts        D-Bus surface typings (Variant, BluezHelper,
                 PropsChangedHandler), helperOf, getDbusNext
* connection.ts  persistent D-Bus connection + adapter singleton,
                 isStale / isDbusConnectionError, dbusError,
                 parseHciIndex
* discovery.ts   startDiscoverySafe (6-tier recovery), removeDevice,
                 autoDiscover, stopDiscoveryAndQuiesce
* freshness.ts   PeerFreshnessTracker, startPeerFreshnessTracker,
                 isPeerFresh
* connect.ts     connectWithRecovery + ConnectRecoveryContext
* gatt.ts        wrapChar, wrapDevice, buildCharMap
* broadcast.ts   extractDbusBytes, broadcastScanNodeBle
* scan.ts        scanAndReadRaw, scanAndRead, scanDevices
* index.ts       public API re-exports + _internals test hooks

persistentConn / persistentAdapter stay module-private in connection.ts
behind exported accessors (getAdapter, resetConnection); the rest of
the dir touches them only through those helpers, never the raw state.

Five import sites updated to handler-node-ble/index.js (Node ESM does
not implicit-resolve directory imports): src/ble/index.ts dynamic
import + scanDevices import, three test files (rssi-freshness,
handler-node-ble-grace, connect-recovery-rssi-skip). CONTRIBUTING.md
file-system map row updated.

No behaviour change. Identical log output, recovery tier sequence,
RSSI freshness logic, GATT timeouts, retry counts, final BlueZ
cleanup sequence (force StopDiscovery + RemoveDevice + btmgmt
power-cycle). 1387/1387 tests pass, lint and tsc clean.
